# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TEST API",
        "description": "TEST API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get posts",
                "operationId": "getAllPosts",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create a new post",
                "operationId": "makePost",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "post_author": {
                                        "type": "integer"
                                    },
                                    "post_content": {
                                        "type": "string"
                                    },
                                    "post_title": {
                                        "type": "string"
                                    },
                                    "post_excerpt": {
                                        "type": "string"
                                    },
                                    "post_status": {
                                        "type": "string"
                                    },
                                    "post_type": {
                                        "type": "string"
                                    },
                                    "post_mime_type": {
                                        "type": "string"
                                    },
                                    "post_guid": {
                                        "type": "string"
                                    },
                                    "post_img": {
                                        "type": "string"
                                    },
                                    "job_salary": {
                                        "type": "string"
                                    },
                                    "job_place": {
                                        "type": "string"
                                    },
                                    "post_tags": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post created successfully"
                    }
                }
            }
        },
        "/api/resource": {
            "get": {
                "tags": [
                    "Resource"
                ],
                "summary": "Get a resource",
                "operationId": "getResource",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    }
}